backend/requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.7
Deprecated==1.2.15
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.9.2
Flask-SQLAlchemy==3.1.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.14.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
psycopg2-binary==2.9.10
Pygments==2.18.0
PyJWT==2.10.1
python-dotenv==1.0.1
rich==13.9.4
SQLAlchemy==2.0.36
typing_extensions==4.12.2
uvicorn==0.32.1
Werkzeug==3.1.3
wrapt==1.17.0
firebase-admin==6.1.0